Feature/add pause

+++ client/helper/receiveFile.py
@@ -2,7 +2,12 @@ import socket
 import os
 import tqdm
 import shutil
+from helper.packetOffet import get_packet_for_filename
+from helper.jsonFileWrite import write_error_data_to_json
+from helper.remove_filename import remove_filename
+import json
 
+json_file_path = r"client\receiveFileData.json"
 BUFFER_SIZE = 1024
 SEPARATOR = "<SEPARATOR>"
 
@@ -40,6 +45,8 @@ def receive_single_file(client_conn, file_info_data):
     try:
         filename, filesize = file_info_data.split(SEPARATOR)
         filename = os.path.basename(filename)
+        basefile_name = filename
+        packet_offset = int(get_packet_for_filename(filename))
         filesize = int(filesize)
 
         progress = tqdm.tqdm(
@@ -51,18 +58,50 @@ def receive_single_file(client_conn, file_info_data):
         )
         downloads_path = get_downloads_folder()
         filename = os.path.join(downloads_path, filename)
-        with open(filename, "wb") as f:
-            while True:
-                bytes_read = client_conn.recv(BUFFER_SIZE)
-                if not bytes_read:
-                    progress.close()
-                    f.close()
-                    break
-                f.write(bytes_read)
-                progress.update(len(bytes_read))
+        bytes_downlaoded = packet_offset
+
+        if os.path.exists(filename):
+            with open(filename, "ab") as f:
+                while True:
+                    bytes_read = client_conn.recv(BUFFER_SIZE)
+                    bytes_downlaoded += len(bytes_read)
+                    if not bytes_read:
+                        progress.close()
+                        f.close()
+                        break
+                    f.write(bytes_read)
+                    progress.update(len(bytes_read))
+
+        else:
+            with open(filename, "wb") as f:
+                while True:
+                    bytes_read = client_conn.recv(BUFFER_SIZE)
+                    bytes_downlaoded += len(bytes_read)
+                    if not bytes_read:
+                        progress.close()
+                        f.close()
+                        break
+                    f.write(bytes_read)
+                    progress.update(len(bytes_read))
+
+        # remove file_name from json file when downloading is completed
+        remove_filename(basefile_name)
+
+        # with open(filename, "wb") as f:
+        #     while True:
+        #         bytes_read = client_conn.recv(BUFFER_SIZE)
+        #         if not bytes_read:
+        #             progress.close()
+        #             f.close()
+        #             break
+        #         f.write(bytes_read)
+        #         progress.update(len(bytes_read))
 
     except Exception as e:
         print(f"Error during file reception: {e}")
+        error_data = {basefile_name: bytes_downlaoded}
+        write_error_data_to_json(error_data)
+
     finally:
         # Close the connection after completing the file/folder reception
         client_conn.close()
@@ -71,28 +110,61 @@ def receive_single_file(client_conn, file_info_data):
 def receive_folder(client_conn, zip_file_path):
     try:
         folder_name = os.path.basename(zip_file_path).replace(".zip", "")
+        basefolder_name = folder_name
+        print("basefolder_name", basefolder_name)
         downloads_path = get_downloads_folder()
         zip_file_path = os.path.join(downloads_path, zip_file_path)
-        receive_zip_file(client_conn, zip_file_path)
+        packet_offset = int(get_packet_for_filename(basefolder_name))
+        bytes_downlaoded = packet_offset
+
+        print("folder_name", basefolder_name)
+        print("exists karta h ki nhi", os.path.exists(zip_file_path))
+
+        if os.path.exists(zip_file_path):
+            with open(zip_file_path, "ab") as f:
+                while True:
+                    bytes_read = client_conn.recv(BUFFER_SIZE)
+                    bytes_downlaoded += len(bytes_read)
+                    if not bytes_read:
+                        break
+                    f.write(bytes_read)
+
+        else:
+            with open(zip_file_path, "wb") as f:
+                while True:
+                    bytes_read = client_conn.recv(BUFFER_SIZE)
+                    bytes_downlaoded += len(bytes_read)
+                    if not bytes_read:
+                        # progress.close()
+                        f.close()
+                        break
+                    f.write(bytes_read)
+
+        # remove folder_name from json file when downloading is completed
+        remove_filename(basefolder_name)
+
+        # receive_zip_file(client_conn, zip_file_path)
         folder_name = os.path.join(downloads_path, folder_name)
         shutil.unpack_archive(zip_file_path, folder_name, "zip")
         os.remove(zip_file_path)
 
     except Exception as e:
         print(f"Error during folder reception: {e}")
+        error_data = {basefolder_name: bytes_downlaoded}
+        write_error_data_to_json(error_data)
 
 
 # new helper function for receiving zip files
-def receive_zip_file(client_conn, zip_file_path):
-    try:
-        with open(zip_file_path, "wb") as f:
-            while True:
-                bytes_read = client_conn.recv(BUFFER_SIZE)
-                if not bytes_read:
-                    break
-                f.write(bytes_read)
-        print(f"File received successfully: {zip_file_path}")
-        return True
-    except Exception as e:
-        print(f"Error receiving file: {str(e)}")
-        return False
+# def receive_zip_file(client_conn, zip_file_path):
+#     try:
+#         with open(zip_file_path, "wb") as f:
+#             while True:
+#                 bytes_read = client_conn.recv(BUFFER_SIZE)
+#                 if not bytes_read:
+#                     break
+#                 f.write(bytes_read)
+#         print(f"File received successfully: {zip_file_path}")
+#         return True
+#     except Exception as e:
+#         print(f"Error receiving file: {str(e)}")
+#         return False
